+++ src/examples/pong/playerSystem.h
@@ -12,5 +12,5 @@ class PlayerSystem : public System
   PlayerSystem();
   ~PlayerSystem();
 
-  void Update(double dt, Registry* registry, Input* input);
+  void Update(float dt, Registry* registry, Input* input);
 };
