+++ src/systems/lucidSystem.h
@@ -16,5 +16,5 @@ class LucidSystem : public System
   LucidSystem();
   ~LucidSystem();
 
-  void Update(double dt, Registry* registry, Input* input);
+  void Update(float dt, Registry* registry, Input* input);
 };
